+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Change Log
 
 ## 1.0.0 (TBD)
 - Migrated from Swagger to OpenApi generator (https://github.com/OpenAPITools/openapi-generator).
+- Empty data class is defined as a String alias.
+- Fixed number enums generated with String value.
+- Changed anonymous enum names from _* to NUMBER_*.
 
 ## 1.0.0 (2019-06-25)
 - Initial version