+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 Change Log
 ==========
 
-## 1.0.0 (TBD)
+## 1.1.0 (TBD)
 - Migrated from Swagger to OpenApi generator (https://github.com/OpenAPITools/openapi-generator).
+- Merged Retrofit and Android generator together.
 - Empty data class is defined as a String alias.
-- Fixed number enums generated with String value.
 - Changed anonymous enum names from _* to NUMBER_*.
 
 ## 1.0.0 (2019-06-25)